Revert "[IN-150] Generalize Providers API"
