+++ api/base/urls.py
@@ -39,7 +39,6 @@ urlpatterns = [
                 url(r'^nodes/', include('api.nodes.urls', namespace='nodes')),
                 url(r'^preprints/', include('api.preprints.urls', namespace='preprints')),
                 url(r'^preprint_providers/', include('api.preprint_providers.urls', namespace='preprint_providers')),
-                url(r'^providers/', include('api.providers.urls', namespace='providers')),
                 url(r'^registrations/', include('api.registrations.urls', namespace='registrations')),
                 url(r'^requests/', include('api.requests.urls', namespace='requests')),
                 url(r'^search/', include('api.search.urls', namespace='search')),
