+++ api/base/views.py
@@ -595,42 +595,3 @@ class WaterButlerMixin(object):
             if check_object_permissions:
                 self.check_object_permissions(self.request, obj)
         return obj
-
-
-class DeprecatedView(JSONAPIBaseView):
-    """ Mixin for deprecating old views
-    Subclasses must define `max_version`
-    """
-
-    @property
-    def max_version(self):
-        raise NotImplementedError()
-
-    def __init__(self, *args, **kwargs):
-        super(DeprecatedView, self).__init__(*args, **kwargs)
-        self.is_deprecated = False
-
-    def determine_version(self, request, *args, **kwargs):
-        version, scheme = super(DeprecatedView, self).determine_version(request, *args, **kwargs)
-        if version > self.max_version:
-            self.is_deprecated = True
-            raise NotFound(detail='This route has been deprecated. It was last available in version {}'.format(self.max_version))
-        return version, scheme
-
-    def finalize_response(self, request, response, *args, **kwargs):
-        response = super(DeprecatedView, self).finalize_response(request, response, *args, **kwargs)
-        if self.is_deprecated:
-            # Already has the error message
-            return response
-        if response.status_code == 204:
-            response.status_code = 200
-            response.data = {}
-        deprecation_warning = 'This route is deprecated and will be unavailable after version {}'.format(self.max_version)
-        if response.data.get('meta', False):
-            if response.data['meta'].get('warnings', False):
-                response.data['meta']['warnings'].append(deprecation_warning)
-            else:
-                response.data['meta']['warnings'] = [deprecation_warning]
-        else:
-            response.data['meta'] = {'warnings': [deprecation_warning]}
-        return response
