+++ api_tests/preprint_providers/views/test_preprint_provider_moderator_list.py
@@ -2,7 +2,7 @@ import mock
 import pytest
 
 from api.base.settings.defaults import API_BASE
-from api.providers.permissions import GroupHelper
+from api.preprint_providers.permissions import GroupHelper
 from osf_tests.factories import (
     AuthUserFactory,
     PreprintProviderFactory,
@@ -18,10 +18,9 @@ class TestPreprintProviderModeratorList:
         GroupHelper(pp).update_provider_auth_groups()
         return pp
 
-    @pytest.fixture(params=['/{}preprint_providers/{}/moderators/', '/{}providers/preprints/{}/moderators/'])
-    def url(self, provider, request):
-        url = request.param
-        return url.format(API_BASE, provider._id)
+    @pytest.fixture()
+    def url(self, provider):
+        return '/{}preprint_providers/{}/moderators/'.format(API_BASE, provider._id)
 
     @pytest.fixture()
     def admin(self, provider):
