+++ api_tests/preprint_providers/views/test_preprint_providers_list.py
@@ -9,10 +9,3 @@ class TestReviewProviderFilters(ReviewProviderFilterMixin):
     @pytest.fixture()
     def url(self):
         return '/{}preprint_providers/'.format(API_BASE)
-
-
-class TestReviewProviderFiltersForGeneralizedEndpoint(ReviewProviderFilterMixin):
-
-    @pytest.fixture()
-    def url(self):
-        return '/{}providers/preprints/'.format(API_BASE)
