+++ api_tests/preprints/filters/test_filters.py
@@ -1,6 +1,6 @@
 import pytest
 
-from api.providers.permissions import GroupHelper
+from api.preprint_providers.permissions import GroupHelper
 from osf_tests.factories import (
     PreprintFactory,
     AuthUserFactory,
