+++ api_tests/reviews/mixins/filter_mixins.py
@@ -3,7 +3,7 @@ from datetime import timedelta
 import pytest
 from furl import furl
 
-from api.providers.permissions import GroupHelper
+from api.preprint_providers.permissions import GroupHelper
 from osf_tests.factories import (
     ReviewActionFactory,
     AuthUserFactory,
