+++ api_tests/users/views/test_user_actions.py
@@ -2,7 +2,7 @@ import pytest
 import mock
 
 from api.base.settings.defaults import API_BASE
-from api.providers.permissions import GroupHelper
+from api.preprint_providers.permissions import GroupHelper
 from osf_tests.factories import (
     PreprintFactory,
     AuthUserFactory,
