+++ osf/models/__init__.py
@@ -22,7 +22,7 @@ from osf.models.private_link import PrivateLink  # noqa
 from osf.models.notifications import NotificationDigest, NotificationSubscription  # noqa
 from osf.models.spam import SpamStatus, SpamMixin  # noqa
 from osf.models.subject import Subject  # noqa
-from osf.models.provider import AbstractProvider, CollectionProvider, PreprintProvider  # noqa
+from osf.models.provider import AbstractProvider, PreprintProvider  # noqa
 from osf.models.preprint_service import PreprintService  # noqa
 from osf.models.request import NodeRequest  # noqa
 from osf.models.identifiers import Identifier  # noqa
