+++ osf/models/provider.py
@@ -8,7 +8,7 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 from dirtyfields import DirtyFieldsMixin
 
-from api.providers.permissions import GroupHelper, PERMISSIONS, GROUP_FORMAT, GROUPS
+from api.preprint_providers.permissions import GroupHelper, PERMISSIONS, GROUP_FORMAT, GROUPS
 from osf.models.base import BaseModel, ObjectIDMixin
 from osf.models.licenses import NodeLicense
 from osf.models.mixins import ReviewProviderMixin
@@ -47,7 +47,8 @@ class AbstractProvider(TypedModel, ObjectIDMixin, ReviewProviderMixin, DirtyFiel
 
     @property
     def all_subjects(self):
-        return self.subjects.all()
+        if self.subjects.exists():
+            return self.subjects.all()
 
     @property
     def has_highlighted_subjects(self):
@@ -66,14 +67,7 @@ class AbstractProvider(TypedModel, ObjectIDMixin, ReviewProviderMixin, DirtyFiel
             return optimize_subject_query(self.subjects.filter(parent__isnull=True))
 
 class CollectionProvider(AbstractProvider):
-    def get_absolute_url(self):
-        return self.absolute_api_v2_url
-
-    @property
-    def absolute_api_v2_url(self):
-        path = '/providers/collections/{}/'.format(self._id)
-        return api_v2_url(path)
-
+    pass
 
 class PreprintProvider(AbstractProvider):
 
@@ -147,7 +141,7 @@ class PreprintProvider(AbstractProvider):
 
     @property
     def absolute_api_v2_url(self):
-        path = '/providers/preprints/{}/'.format(self._id)
+        path = '/preprint_providers/{}/'.format(self._id)
         return api_v2_url(path)
 
     def save(self, *args, **kwargs):
